+++ index.js
@@ -133,13 +133,12 @@ module.exports =
 	            btnDownActive: false,
 	            btnUpHover: false,
 	            btnUpActive: false,
-	            inputFocus: false,
-
 	            stringValue: ""
 	        }, _this._propsToState(_this.props));
-
 	        _this.stop = _this.stop.bind(_this);
 	        _this.onTouchEnd = _this.onTouchEnd.bind(_this);
+	        _this.refsInput = {};
+	        _this.refsWrapper = {};
 	        return _this;
 	    }
 
@@ -183,18 +182,18 @@ module.exports =
 	        key: 'componentDidUpdate',
 	        value: function componentDidUpdate(prevProps, prevState) {
 	            if (!this._ignoreValueChange && prevState.value !== this.state.value && (!isNaN(this.state.value) || this.state.value === null)) {
-	                    this._invokeEventCallback("onChange", this.state.value, this.refs.input.value, this.refs.input);
+	                    this._invokeEventCallback("onChange", this.state.value, this.refsInput.value, this.refsInput);
 	                }
 
-	            if (this.state.inputFocus) {
-	                this.refs.input.focus();
+	            if (this._inputFocus) {
+	                this.refsInput.focus();
 
 	                if (this.state.selectionStart || this.state.selectionStart === 0) {
-	                    this.refs.input.selectionStart = this.state.selectionStart;
+	                    this.refsInput.selectionStart = this.state.selectionStart;
 	                }
 
 	                if (this.state.selectionEnd || this.state.selectionEnd === 0) {
-	                    this.refs.input.selectionEnd = this.state.selectionEnd;
+	                    this.refsInput.selectionEnd = this.state.selectionEnd;
 	                }
 	            }
 
@@ -212,22 +211,22 @@ module.exports =
 	            var _this3 = this;
 
 	            this._isMounted = true;
-	            this.refs.input.getValueAsNumber = function () {
+	            this.refsInput.getValueAsNumber = function () {
 	                return _this3.state.value || 0;
 	            };
 
-	            this.refs.input.setValue = function (value) {
+	            this.refsInput.setValue = function (value) {
 	                _this3.setState({
 	                    value: _this3._parse(value),
 	                    stringValue: value
 	                });
 	            };
 
-	            if (!this.state.inputFocus && IS_BROWSER && document.activeElement === this.refs.input) {
-	                this.state.inputFocus = true;
-	                this.refs.input.focus();
+	            if (!this._inputFocus && IS_BROWSER && document.activeElement === this.refsInput) {
+	                this._inputFocus = true;
+	                this.refsInput.focus();
 	                this._invokeEventCallback("onFocus", {
-	                    target: this.refs.input,
+	                    target: this.refsInput,
 	                    type: "focus"
 	                });
 	            }
@@ -237,8 +236,8 @@ module.exports =
 	    }, {
 	        key: 'saveSelection',
 	        value: function saveSelection() {
-	            this.state.selectionStart = this.refs.input.selectionStart;
-	            this.state.selectionEnd = this.refs.input.selectionEnd;
+	            this.state.selectionStart = this.refsInput.selectionStart;
+	            this.state.selectionEnd = this.refsInput.selectionEnd;
 	        }
 	    }, {
 	        key: 'checkValidity',
@@ -246,50 +245,50 @@ module.exports =
 	            var valid = void 0,
 	                validationError = "";
 
-	            var supportsValidation = !!this.refs.input.checkValidity;
+	            var supportsValidation = !!this.refsInput.checkValidity;
 
 	            var noValidate = !!(this.props.noValidate && this.props.noValidate != "false");
 
-	            this.refs.input.noValidate = noValidate;
+	            this.refsInput.noValidate = noValidate;
 
 	            valid = noValidate || !supportsValidation;
 
 	            if (valid) {
 	                validationError = "";
 	            } else {
-	                if (this.refs.input.pattern === "") {
-	                    this.refs.input.pattern = this.props.required ? ".+" : ".*";
+	                if (this.refsInput.pattern === "") {
+	                    this.refsInput.pattern = this.props.required ? ".+" : ".*";
 	                }
 
 	                if (supportsValidation) {
-	                    this.refs.input.checkValidity();
-	                    valid = this.refs.input.validity.valid;
+	                    this.refsInput.checkValidity();
+	                    valid = this.refsInput.validity.valid;
 
 	                    if (!valid) {
-	                        validationError = this.refs.input.validationMessage;
+	                        validationError = this.refsInput.validationMessage;
 	                    }
 	                }
 
 	                if (valid && supportsValidation && this.props.maxLength) {
-	                    if (this.refs.input.value.length > this.props.maxLength) {
+	                    if (this.refsInput.value.length > this.props.maxLength) {
 	                        validationError = "This value is too long";
 	                    }
 	                }
 	            }
 
-	            validationError = validationError || (valid ? "" : this.refs.input.validationMessage || "Unknown Error");
+	            validationError = validationError || (valid ? "" : this.refsInput.validationMessage || "Unknown Error");
 
 	            var validStateChanged = this._valid !== validationError;
 	            this._valid = validationError;
 	            if (validationError) {
-	                addClass(this.refs.wrapper, "has-error");
+	                addClass(this.refsWrapper, "has-error");
 	                if (validStateChanged) {
-	                    this._invokeEventCallback("onInvalid", validationError, this.state.value, this.refs.input.value);
+	                    this._invokeEventCallback("onInvalid", validationError, this.state.value, this.refsInput.value);
 	                }
 	            } else {
-	                removeClass(this.refs.wrapper, "has-error");
+	                removeClass(this.refsWrapper, "has-error");
 	                if (validStateChanged) {
-	                    this._invokeEventCallback("onValid", this.state.value, this.refs.input.value);
+	                    this._invokeEventCallback("onValid", this.state.value, this.refsInput.value);
 	                }
 	            }
 	        }
@@ -379,17 +378,17 @@ module.exports =
 	                    e.preventDefault();
 	                    this._step(e.ctrlKey || e.metaKey ? -0.1 : e.shiftKey ? -10 : -1);
 	                } else {
-	                    var _value = this.refs.input.value,
+	                    var _value = this.refsInput.value,
 	                        length = _value.length;
 	                    if (e.keyCode === 8) {
-	                        if (this.refs.input.selectionStart == this.refs.input.selectionEnd && this.refs.input.selectionEnd > 0 && _value.length && _value.charAt(this.refs.input.selectionEnd - 1) === ".") {
+	                        if (this.refsInput.selectionStart == this.refsInput.selectionEnd && this.refsInput.selectionEnd > 0 && _value.length && _value.charAt(this.refsInput.selectionEnd - 1) === ".") {
 	                            e.preventDefault();
-	                            this.refs.input.selectionStart = this.refs.input.selectionEnd = this.refs.input.selectionEnd - 1;
+	                            this.refsInput.selectionStart = this.refsInput.selectionEnd = this.refsInput.selectionEnd - 1;
 	                        }
 	                    } else if (e.keyCode === 46) {
-	                        if (this.refs.input.selectionStart == this.refs.input.selectionEnd && this.refs.input.selectionEnd < length + 1 && _value.length && _value.charAt(this.refs.input.selectionEnd) === ".") {
+	                        if (this.refsInput.selectionStart == this.refsInput.selectionEnd && this.refsInput.selectionEnd < length + 1 && _value.length && _value.charAt(this.refsInput.selectionEnd) === ".") {
 	                            e.preventDefault();
-	                            this.refs.input.selectionStart = this.refs.input.selectionEnd = this.refs.input.selectionEnd + 1;
+	                            this.refsInput.selectionStart = this.refsInput.selectionEnd = this.refsInput.selectionEnd + 1;
 	                        }
 	                    } else if(!validInputKey(e.keyCode)) {
 			    	e.preventDefault(); // don't allow non-numberic characters
@@ -496,6 +495,7 @@ module.exports =
 	                format = _props.format,
 	                mobile = _props.mobile,
 	                snap = _props.snap,
+	                componentClass = _props.componentClass,
 	                value = _props.value,
 	                type = _props.type,
 	                style = _props.style,
@@ -503,7 +503,10 @@ module.exports =
 	                onInvalid = _props.onInvalid,
 	                onValid = _props.onValid,
 	                strict = _props.strict,
-	                rest = _objectWithoutProperties(_props, ['step', 'min', 'max', 'precision', 'parse', 'format', 'mobile', 'snap', 'value', 'type', 'style', 'defaultValue', 'onInvalid', 'onValid', 'strict']);
+	                noStyle = _props.noStyle,
+	                rest = _objectWithoutProperties(_props, ['step', 'min', 'max', 'precision', 'parse', 'format', 'mobile', 'snap', 'componentClass', 'value', 'type', 'style', 'defaultValue', 'onInvalid', 'onValid', 'strict', 'noStyle']);
+
+	            noStyle = noStyle || style === false;
 
 	            for (var x in NumericInput.style) {
 	                css[x] = _extends({}, NumericInput.style[x], style ? style[x] || {} : {});
@@ -522,16 +525,24 @@ module.exports =
 
 	            var attrs = {
 	                wrap: {
-	                    style: style === false ? null : css.wrap,
+	                    style: noStyle ? null : css.wrap,
 	                    className: 'react-numeric-input',
-	                    ref: 'wrapper',
+	                    ref: function ref(e) {
+	                        if (e != null && e != undefined) {
+	                            _this6.refsWrapper = e;
+	                        }
+	                    },
 	                    onMouseUp: undefined,
 	                    onMouseLeave: undefined
 	                },
 	                input: _extends({
-	                    ref: 'input',
+	                    ref: function ref(e) {
+	                        if (e != null && e != undefined) {
+	                            _this6.refsInput = e;
+	                        }
+	                    },
 	                    type: 'text',
-	                    style: style === false ? null : _extends({}, css.input, !hasFormControl ? css['input:not(.form-control)'] : {}, state.inputFocus ? css['input:focus'] : {})
+	                    style: noStyle ? null : _extends({}, css.input, !hasFormControl ? css['input:not(.form-control)'] : {}, this._inputFocus ? css['input:focus'] : {})
 	                }, rest),
 	                btnUp: {
 	                    onMouseEnter: undefined,
@@ -540,7 +551,7 @@ module.exports =
 	                    onMouseLeave: undefined,
 	                    onTouchStart: undefined,
 	                    onTouchEnd: undefined,
-	                    style: style === false ? null : _extends({}, css.btn, css.btnUp, props.disabled ? css['btn:disabled'] : state.btnUpActive ? css['btn:active'] : state.btnUpHover ? css['btn:hover'] : {})
+	                    style: noStyle ? null : _extends({}, css.btn, css.btnUp, props.disabled || props.readOnly ? css['btn:disabled'] : state.btnUpActive ? css['btn:active'] : state.btnUpHover ? css['btn:hover'] : {})
 	                },
 	                btnDown: {
 	                    onMouseEnter: undefined,
@@ -549,15 +560,17 @@ module.exports =
 	                    onMouseLeave: undefined,
 	                    onTouchStart: undefined,
 	                    onTouchEnd: undefined,
-	                    style: style === false ? null : _extends({}, css.btn, css.btnDown, props.disabled ? css['btn:disabled'] : state.btnDownActive ? css['btn:active'] : state.btnDownHover ? css['btn:hover'] : {})
+	                    style: noStyle ? null : _extends({}, css.btn, css.btnDown, props.disabled || props.readOnly ? css['btn:disabled'] : state.btnDownActive ? css['btn:active'] : state.btnDownHover ? css['btn:hover'] : {})
 	                }
 	            };
 
 	            var stringValue = String(state.stringValue || (state.value || state.value === 0 ? state.value : "") || "");
 
-	            if (RE_INCOMPLETE_NUMBER.test(stringValue)) {
+	            var loose = !this._isStrict && (this._inputFocus || !this._isMounted);
+
+	            if (loose && RE_INCOMPLETE_NUMBER.test(stringValue)) {
 	                attrs.input.value = stringValue;
-	            } else if (!this._isStrict && stringValue && !RE_NUMBER.test(stringValue)) {
+	            } else if (loose && stringValue && !RE_NUMBER.test(stringValue)) {
 	                    attrs.input.value = stringValue;
 	                } else if (state.value || state.value === 0) {
 	                        attrs.input.value = this._format(state.value);
@@ -565,17 +578,17 @@ module.exports =
 	                            attrs.input.value = "";
 	                        }
 
-	            if (hasFormControl && style !== false) {
+	            if (hasFormControl && !noStyle) {
 	                _extends(attrs.wrap.style, css['wrap.hasFormControl']);
 	            }
 
-	            if (mobile && style !== false) {
+	            if (mobile && !noStyle) {
 	                _extends(attrs.input.style, css['input.mobile']);
 	                _extends(attrs.btnUp.style, css['btnUp.mobile']);
 	                _extends(attrs.btnDown.style, css['btnDown.mobile']);
 	            }
 
-	            if (!props.disabled) {
+	            if (!props.disabled && !props.readOnly) {
 	                _extends(attrs.wrap, {
 	                    onMouseUp: this.stop,
 	                    onMouseLeave: this.stop
@@ -609,10 +622,10 @@ module.exports =
 
 	                        args[0].preventDefault();
 	                        args[0].persist();
+	                        _this6._inputFocus = true;
 	                        _this6.setState({
 	                            btnUpHover: true,
-	                            btnUpActive: true,
-	                            inputFocus: true
+	                            btnUpActive: true
 	                        }, function () {
 	                            _this6._invokeEventCallback.apply(_this6, ["onFocus"].concat(args));
 	                            _this6.onMouseDown('up');
@@ -648,10 +661,10 @@ module.exports =
 
 	                        args[0].preventDefault();
 	                        args[0].persist();
+	                        _this6._inputFocus = true;
 	                        _this6.setState({
 	                            btnDownHover: true,
-	                            btnDownActive: true,
-	                            inputFocus: true
+	                            btnDownActive: true
 	                        }, function () {
 	                            _this6._invokeEventCallback.apply(_this6, ["onFocus"].concat(args));
 	                            _this6.onMouseDown('down');
@@ -694,14 +707,13 @@ module.exports =
 	                        }
 
 	                        args[0].persist();
-	                        _this6.setState({ inputFocus: true }, function () {
-	                            var val = _this6._parse(args[0].target.value);
-	                            _this6.setState({
-	                                value: val,
-	                                stringValue: val || val === 0 ? val + "" : ""
-	                            }, function () {
-	                                _this6._invokeEventCallback.apply(_this6, ["onFocus"].concat(args));
-	                            });
+	                        _this6._inputFocus = true;
+	                        var val = _this6._parse(args[0].target.value);
+	                        _this6.setState({
+	                            value: val,
+	                            stringValue: val || val === 0 ? val + "" : ""
+	                        }, function () {
+	                            _this6._invokeEventCallback.apply(_this6, ["onFocus"].concat(args));
 	                        });
 	                    },
 	                    onBlur: function onBlur() {
@@ -712,38 +724,39 @@ module.exports =
 	                        var _isStrict = _this6._isStrict;
 	                        _this6._isStrict = true;
 	                        args[0].persist();
-	                        _this6.setState({ inputFocus: false }, function () {
-	                            var val = _this6._parse(args[0].target.value);
-	                            _this6.setState({
-	                                value: val
-	                            }, function () {
-	                                _this6._invokeEventCallback.apply(_this6, ["onBlur"].concat(args));
-	                                _this6._isStrict = _isStrict;
-	                            });
+	                        _this6._inputFocus = false;
+	                        var val = _this6._parse(args[0].target.value);
+	                        _this6.setState({
+	                            value: val
+	                        }, function () {
+	                            _this6._invokeEventCallback.apply(_this6, ["onBlur"].concat(args));
+	                            _this6._isStrict = _isStrict;
 	                        });
 	                    }
 	                });
 	            } else {
-	                if (style !== false) {
+	                if (!noStyle && props.disabled) {
 	                    _extends(attrs.input.style, css['input:disabled']);
 	                }
 	            }
+
+	            var InputTag = componentClass || 'input';
 
 	            if (mobile) {
 	                return _react2.default.createElement(
 	                    'span',
 	                    attrs.wrap,
-	                    _react2.default.createElement('input', attrs.input),
+	                    _react2.default.createElement(InputTag, attrs.input),
 	                    _react2.default.createElement(
 	                        'b',
 	                        attrs.btnUp,
-	                        _react2.default.createElement('i', { style: style === false ? null : css.minus }),
-	                        _react2.default.createElement('i', { style: style === false ? null : css.plus })
+	                        _react2.default.createElement('i', { style: noStyle ? null : css.minus }),
+	                        _react2.default.createElement('i', { style: noStyle ? null : css.plus })
 	                    ),
 	                    _react2.default.createElement(
 	                        'b',
 	                        attrs.btnDown,
-	                        _react2.default.createElement('i', { style: style === false ? null : css.minus })
+	                        _react2.default.createElement('i', { style: noStyle ? null : css.minus })
 	                    )
 	                );
 	            }
@@ -751,16 +764,16 @@ module.exports =
 	            return _react2.default.createElement(
 	                'span',
 	                attrs.wrap,
-	                _react2.default.createElement('input', attrs.input),
+	                _react2.default.createElement(InputTag, attrs.input),
 	                _react2.default.createElement(
 	                    'b',
 	                    attrs.btnUp,
-	                    _react2.default.createElement('i', { style: style === false ? null : css.arrowUp })
+	                    _react2.default.createElement('i', { style: noStyle ? null : css.arrowUp })
 	                ),
 	                _react2.default.createElement(
 	                    'b',
 	                    attrs.btnDown,
-	                    _react2.default.createElement('i', { style: style === false ? null : css.arrowDown })
+	                    _react2.default.createElement('i', { style: noStyle ? null : css.arrowDown })
 	                )
 	            );
 	        }
@@ -784,6 +797,7 @@ module.exports =
 	    snap: _propTypes2.default.bool,
 	    noValidate: _propTypes2.default.oneOfType([_propTypes2.default.bool, _propTypes2.default.string]),
 	    style: _propTypes2.default.oneOfType([_propTypes2.default.object, _propTypes2.default.bool]),
+	    noStyle: _propTypes2.default.bool,
 	    type: _propTypes2.default.string,
 	    pattern: _propTypes2.default.string,
 	    onFocus: _propTypes2.default.func,
@@ -798,6 +812,7 @@ module.exports =
 	    value: _propTypes2.default.oneOfType([_propTypes2.default.number, _propTypes2.default.string]),
 	    defaultValue: _propTypes2.default.oneOfType([_propTypes2.default.number, _propTypes2.default.string]),
 	    strict: _propTypes2.default.bool,
+	    componentClass: _propTypes2.default.string,
 	    mobile: function mobile(props, propName) {
 	        var prop = props[propName];
 	        if (prop !== true && prop !== false && prop !== 'auto' && typeof prop != 'function') {
@@ -814,6 +829,7 @@ module.exports =
 	    format: null,
 	    mobile: 'auto',
 	    strict: false,
+	    componentClass: "input",
 	    style: {}
 	};
 	NumericInput.style = {
@@ -933,7 +949,8 @@ module.exports =
 
 	    input: {
 	        paddingRight: '3ex',
-	        boxSizing: 'border-box'
+	        boxSizing: 'border-box',
+	        fontSize: 'inherit'
 	    },
 
 	    'input:not(.form-control)': {